Added Metal cubemap textures, no data yet

+++ Hazard-Renderer/src/Backend/OpenGL/OpenGLRenderCommandBuffer.h
@@ -31,8 +31,8 @@ namespace HazardRenderer::OpenGL
 		void BindUniformBuffer(Ref<UniformBuffer> uniformBuffer) override;
 		void BindPipeline(Ref<Pipeline> pipeline) override;
 
-		void Draw(uint32_t count, Ref<IndexBuffer> indexBuffer = nullptr) override;
-		void DrawInstanced(uint32_t count, uint32_t instanceCount, Ref<IndexBuffer> indexBuffer = nullptr) override;
+		void Draw(size_t count, Ref<IndexBuffer> indexBuffer = nullptr) override;
+		void DrawInstanced(size_t count, uint32_t instanceCount, Ref<IndexBuffer> indexBuffer = nullptr) override;
 		void DispatchCompute(const DispatchComputeInfo& computeInfo) override;
 		void TraceRays(const TraceRaysInfo& traceRaysInfo) override {};
 		void BuildAccelerationStructure(const AccelerationStructureBuildInfo& info) {};
